Initial commit of cinder-common
Noticed that every recipe in the cookbook had a declaration for the
cinder.conf template, with a notify to the service defined in that recipie.

I moved the template, and the variables that it needs to populate,
to cinder-common.rb, included that in all the other recipes,
and changed the notify in the template to subscribes in each service
Also, added a spec for this.

All tests pass

Change-Id: Ibcab9566a907e1ea9febbd61e37b9587625a15e6