Only lookup volume service passwords when necessary
The volume service passwords should be looked up depending on
the type of driver used.  As I understand it the grizzly netapp
driver no longer requires dfm.  However, without testing this,
would like to keep the existing configuraton.  We will be testing
this in our lab over the next few days, and will send a new patch
based on our findings.

Change-Id: I1cf0b6e292eec71788d0b79dbf270a1d17bbfa1e