Adding support for netapp NFS direct driver
This commit adds support for the 7mode NFS direct driver.
Changed the 'if driver' to a case, somewhat based off the
rcbops cinder-volume recipe. if the NFS driver is chosen, the
volume recipe will install the nfs client package, setup a mountpoint
a shares config file, and setups cinder.conf accordingly

Change-Id: Ia8d4c31235426614a317729494569062a23a05bf