Updates cinder-group-active script to work on Ubuntu
Setting the following attribute to true causes a chef-client to fail on
Ubuntu:

default['openstack']['block-storage']['volume']['create_volume_group']

The reason is that when this is enabled, the file cinder-group-active
is dropped in /etc/init.d and attempting to start it fails since it's
not written for Ubuntu.  This change wraps some of the Red Hat specific
bits with if statements and also conditionally runs the vgcreate so we
don't run losetup unnecessarily (which keeps mapping the volume file to
new /dev/loop devices).

Change-Id: I5e14aeec64b1c1d596584bfc5d3352f45114421a
Closes-Bug: #1274385