Allows for flexible NFS volume naming
It doesn't make sense to force all NFS exports to have the same
volume name, as you could be mounting from various different sources.
This change allows the user to specify the values for shares.conf
instead of having it rendered from "netapp_server_hostname:@export"

This is explicity for grizzly today, as I need it, but Blueprint is
submitted for allowing more generic NFS cinder in master.

Also adds rubocop as tests fail without it

Change-Id: Icd5c6d92a9ae8f5f68f8bb1ad62d180f388fde90